Add LingoClashDev to target membership
