+++ LingoClash/LingoClash/Model/BookCategory.swift
@@ -10,4 +10,10 @@ import Foundation
 struct BookCategory {
     let id: Identifier
     let name: String
+    
+    init(bookCategoryData: BookCategoryData) {
+        self.id = bookCategoryData.id
+        self.name = bookCategoryData.name
+    }
 }
+
